pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safeai"
version = "0.1.0rc1"
description = "Framework-agnostic runtime boundary enforcement for AI systems"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "SafeAI Contributors" }
]
dependencies = [
  "click>=8.1,<9",
  "pydantic>=2.7,<3",
  "pydantic-settings>=2.2,<3",
  "PyYAML>=6.0,<7",
  "jsonschema>=4.22,<5",
  "structlog>=24,<26",
  "fastapi>=0.110,<1",
  "uvicorn>=0.29,<1",
  "cryptography>=42,<44"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
  "pytest-benchmark>=4.0,<5",
  "ruff>=0.4,<1",
  "mypy>=1.9,<2"
]

[project.scripts]
safeai = "safeai.cli.main:cli"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["safeai*"]

[tool.setuptools.package-data]
"safeai" = ["config/defaults/**/*.yaml", "schemas/**/*.json"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_configs = true
check_untyped_defs = true
disallow_incomplete_defs = false
disallow_untyped_defs = false
no_implicit_optional = true
